Create myLibrary.js for reuseble funcs
