+++ src/components/recipe/RecipeList/RecipeCard/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
-import { Link } from 'react-router-dom'
 import CardReveal from './CardReveal'
+import { trimString } from '../../../../myLibrary'
 
 const RecipeCard = props => {
   const {
@@ -8,12 +8,6 @@ const RecipeCard = props => {
     vegetarian, vegan, dairyFree, readyInMinutes,
     selectRecipe, recipes, cookingMinutes
   } = props
-
-  const trimString = (string, length) => {
-    return string.length > length
-      ? string.substring(0, length -3) + "..."
-      : string
-  }
 
   return (
     <div
@@ -23,7 +17,7 @@ const RecipeCard = props => {
       onClick={() => { selectRecipe(recipes.find(rec => rec.id === id)) }}
       >
       {/*Later => Use Header component and pass style as prop.
-      Make use of default props in Header component */}
+      Make use of default props in Header component if necessary */}
       <div className="card-image waves-effect waves-block waves-light">
         <img className="activator" src={image} alt={title} />
       </div>
@@ -31,8 +25,6 @@ const RecipeCard = props => {
       <span class="card-title activator grey-text text-darken-4">
         { trimString(title, 40) }
       </span>
-        {/* <span className="card-title activator grey-text text-darken-4">Flip</span>
-        <Link to={`/recipes/${id}`}></Link> */}
       </div>
       <CardReveal
         title={title}
